spec: added patch for centos support
This patch adds a CentOS version of the CloudInit class, and changes juju to use this class.

The _collect_packages method returns the name of rpm packages for yum. The render method
invokes the format_cloud_init function providing three modules: ssh, runcmd and yum_packages.

The latter module requires the patched version of cloud-init, that can be found at
https://github.com/timeredbull/cloudinit-centos-6